+++ src/test/java/ru/yandex/market/bazhenava/ui/login/LoginInputIncorrectDateTest.java
@@ -1,22 +1,25 @@
-package ru.yandex.market.bazhenava.login;
+package ru.yandex.market.bazhenava.ui.login;
 
 import org.junit.jupiter.api.Test;
 import ru.yandex.market.bazhenava.HomePage;
 import ru.yandex.market.bazhenava.LoginPage;
 import ru.yandex.market.bazhenava.driver.DriverSingleton;
 
-public class LoginWithEmtyEmailDateTest {
+public class LoginInputIncorrectDateTest {
 
     @Test
-    public void testLoginWithEmptyEmailData() {
-
+    public void testInputIncorrectEmailAndPasswordDate() {
         HomePage homePage = new HomePage();
         homePage.openPage();
         homePage.clickButtonLogin();
 
         LoginPage loginPage = new LoginPage();
+        loginPage.inputEmail("mail");
         loginPage.clickButtonEnter();
-        loginPage.validatingWhenEmptyEmailEnter();
+        loginPage.inputPassword("1234");
+        loginPage.clickButtonLoginContinue();
+
+        loginPage.validatingWhenIncorrectPasswordEnter();
         DriverSingleton.closeDriver();
     }
 }
