+++ src/test/java/ru/yandex/market/bazhenava/ui/login/LoginWithEmtyEmailDateTest.java
@@ -1,24 +1,22 @@
-package ru.yandex.market.bazhenava.login;
+package ru.yandex.market.bazhenava.ui.login;
 
 import org.junit.jupiter.api.Test;
 import ru.yandex.market.bazhenava.HomePage;
 import ru.yandex.market.bazhenava.LoginPage;
 import ru.yandex.market.bazhenava.driver.DriverSingleton;
 
-public class LoginWithEmptyPhoneDataTest {
-
+public class LoginWithEmtyEmailDateTest {
 
     @Test
-    public void testLoginWithEmptyPhoneData() {
+    public void testLoginWithEmptyEmailData() {
 
         HomePage homePage = new HomePage();
         homePage.openPage();
         homePage.clickButtonLogin();
 
         LoginPage loginPage = new LoginPage();
-        loginPage.clickButtonPhoneEnter();
         loginPage.clickButtonEnter();
-        loginPage.validatingWhenEmptyPhoneEnter();
+        loginPage.validatingWhenEmptyEmailEnter();
         DriverSingleton.closeDriver();
     }
 }
